+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### 🐞 Bug fixes
 - Prevent original input style JSON from being mutated by `Style.set*` methods ([#6216](https://github.com/maplibre/maplibre-gl-js/pull/6216))
+- Fix evaluating `global-state` in paint properties with other subexpressions ([#6048](https://github.com/maplibre/maplibre-gl-js/issues/6048))
 
 ## 5.6.2